Add Cms Page Conector, Adapter, Normalizer
