+++ projects/core/src/cms/connectors/index.ts
@@ -1,0 +1,2 @@
+export * from './page/index';
+export * from './component/index';
