+++ projects/core/src/cms/connectors/page/cms-page.connector.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CmsPageConnector } from './cms-page.connector';
+
+describe('CmsPageConnectorService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: CmsPageConnector = TestBed.get(CmsPageConnector);
+    expect(service).toBeTruthy();
+  });
+});
