+++ projects/core/src/cms/connectors/page/cms-page.connector.ts
@@ -1,33 +1,23 @@
-import { HttpErrorResponse } from '@angular/common/http';
-import { Injectable, Optional } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { Observable, of, throwError } from 'rxjs';
-import { catchError, map, switchMap } from 'rxjs/operators';
-import { PageContext } from '../../routing/models/page-context.model';
-import { CmsStructureModel } from '../model/page.model';
+import { catchError, switchMap } from 'rxjs/operators';
+import { HttpErrorResponse } from '@angular/common/http';
 import { CmsPageAdapter } from './cms-page.adapter';
-import { CmsStructureConfigService } from './cms-structure-config.service';
+import { CmsStructureConfigService } from '../../services/cms-structure-config.service';
+import { PageContext } from '../../../routing/models/page-context.model';
+import { CmsStructureModel } from '../../model/page.model';
+import { CMS_PAGE_NORMALIZER } from './cms-page.normalizer';
+import { NormalizersService } from '../../../util/normalizers.service';
 
-/**
- * Abstract class that can be used to implement custom loader logic
- * in order to load CMS structure from third-party CMS system.
- */
 @Injectable({
   providedIn: 'root',
 })
-export abstract class CmsPageLoader<T> {
+export class CmsPageConnector {
   constructor(
+    protected cmsPageAdapter: CmsPageAdapter<any>,
     protected cmsStructureConfigService: CmsStructureConfigService,
-    @Optional() protected adapter: CmsPageAdapter<T>
+    protected normalizers: NormalizersService
   ) {}
-
-  /**
-   * Abstract method must be used to load the page structure for a given `PageContext`.
-   * The page can be loaded from alternative sources, as long as the structure
-   * converts to the `CmsStructureModel`.
-   *
-   * @param pageContext The `PageContext` holding the page Id.
-   */
-  abstract load(pageContext: PageContext): Observable<T>;
 
   /**
    * Returns an observable with the page structure. The page structure is
@@ -40,8 +30,8 @@ export abstract class CmsPageLoader<T> {
       .pipe(
         switchMap(loadFromConfig => {
           if (!loadFromConfig) {
-            return this.load(pageContext).pipe(
-              map(page => this.adapt(page)),
+            return this.cmsPageAdapter.load(pageContext).pipe(
+              this.normalizers.pipeable(CMS_PAGE_NORMALIZER),
               catchError(error => {
                 if (
                   error instanceof HttpErrorResponse &&
@@ -63,20 +53,6 @@ export abstract class CmsPageLoader<T> {
 
   /**
    *
-   * An adapter can be injected to convert the backend reponse to
-   * the UI model.
-   *
-   * @param page the source that can be converted
-   */
-  adapt(page: T): CmsStructureModel {
-    if (this.adapter) {
-      return this.adapter.adapt(<T>page);
-    }
-    return <CmsStructureModel>page;
-  }
-
-  /**
-   *
    * Merge default page structure inot the given `CmsStructureModel`.
    * This is benefitial for a fast setup of the UI without necessary
    * finegrained CMS setup.
@@ -90,4 +66,5 @@ export abstract class CmsPageLoader<T> {
       pageStructure
     );
   }
+
 }
