+++ projects/core/src/cms/connectors/page/index.ts
@@ -1,0 +1,3 @@
+export * from './cms-page.adapter';
+export * from './cms-page.connector';
+export * from './cms-page.normalizer';
