+++ projects/core/src/cms/occ/index.ts
@@ -1,3 +1,2 @@
-export * from './occ-cms-page.loader';
-export * from './occ-cms-page.loader';
 export * from './occ-cms-page.adapter';
+export * from './occ-cms-page.normalizer';
