--- projects/core/src/cms/occ/occ-cms-page.loader.ts
@@ -1,118 +1,0 @@
-import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
-import { Injectable, Optional } from '@angular/core';
-import { Observable, throwError } from 'rxjs';
-import { catchError } from 'rxjs/operators';
-import {
-  CmsComponentList,
-  CMSPage,
-  PageType,
-} from '../../occ/occ-models/index';
-import { OccEndpointsService } from '../../occ/services/occ-endpoints.service';
-import { PageContext } from '../../routing/index';
-import { CmsStructureConfig } from '../config/cms-structure.config';
-import { IdList } from '../model/idList.model';
-import { CmsPageAdapter } from '../services/cms-page.adapter';
-import { CmsPageLoader } from '../services/cms-page.loader';
-import { CmsStructureConfigService } from '../services/cms-structure-config.service';
-
-@Injectable()
-export class OccCmsPageLoader extends CmsPageLoader<CMSPage> {
-  protected headers = new HttpHeaders().set('Content-Type', 'application/json');
-
-  constructor(
-    private http: HttpClient,
-    protected config: CmsStructureConfig,
-    protected cmsStructureConfigService: CmsStructureConfigService,
-    @Optional() protected adapter: CmsPageAdapter<CMSPage>,
-    private occEndpoints: OccEndpointsService
-  ) {
-    super(cmsStructureConfigService, adapter);
-  }
-
-  protected getBaseEndPoint(): string {
-    return this.occEndpoints.getEndpoint('cms');
-  }
-
-  load(pageContext: PageContext, fields?: string): Observable<CMSPage> {
-    let httpStringParams = '';
-
-    if (pageContext.id !== 'smartedit-preview') {
-      httpStringParams = 'pageType=' + pageContext.type;
-
-      if (pageContext.type === PageType.CONTENT_PAGE) {
-        httpStringParams =
-          httpStringParams + '&pageLabelOrId=' + pageContext.id;
-      } else {
-        httpStringParams = httpStringParams + '&code=' + pageContext.id;
-      }
-    }
-
-    if (fields !== undefined) {
-      httpStringParams = httpStringParams + '&fields=' + fields;
-    }
-
-    return this.http.get(this.getBaseEndPoint() + `/pages`, {
-      headers: this.headers,
-      params: new HttpParams({
-        fromString: httpStringParams,
-      }),
-    });
-  }
-
-  loadListComponents(
-    idList: IdList,
-    pageContext: PageContext,
-    fields?: string,
-    currentPage?: number,
-    pageSize?: number,
-    sort?: string
-  ): Observable<CmsComponentList> {
-    let requestParams = this.getRequestParams(pageContext, fields);
-    if (currentPage !== undefined) {
-      requestParams === ''
-        ? (requestParams = requestParams + 'currentPage=' + currentPage)
-        : (requestParams = requestParams + '&currentPage=' + currentPage);
-    }
-    if (pageSize !== undefined) {
-      requestParams = requestParams + '&pageSize=' + pageSize;
-    }
-    if (sort !== undefined) {
-      requestParams = requestParams + '&sort=' + sort;
-    }
-
-    return this.http
-      .post<CmsComponentList>(this.getBaseEndPoint() + `/components`, idList, {
-        headers: this.headers,
-        params: new HttpParams({
-          fromString: requestParams,
-        }),
-      })
-      .pipe(catchError((error: any) => throwError(error.json())));
-  }
-
-  private getRequestParams(pageContext: PageContext, fields?: string): string {
-    let requestParams = '';
-    switch (pageContext.type) {
-      case PageType.PRODUCT_PAGE: {
-        requestParams = 'productCode=' + pageContext.id;
-        break;
-      }
-      case PageType.CATEGORY_PAGE: {
-        requestParams = 'categoryCode=' + pageContext.id;
-        break;
-      }
-      case PageType.CATALOG_PAGE: {
-        requestParams = 'catalogCode=' + pageContext.id;
-        break;
-      }
-    }
-
-    if (fields !== undefined) {
-      requestParams === ''
-        ? (requestParams = requestParams + 'fields=' + fields)
-        : (requestParams = requestParams + '&fields=' + fields);
-    }
-
-    return requestParams;
-  }
-}
