--- projects/core/src/cms/services/cms-page.adapter.ts
@@ -1,7 +1,0 @@
-import { Injectable } from '@angular/core';
-import { CmsStructureModel } from '../model/page.model';
-
-@Injectable()
-export abstract class CmsPageAdapter<S> {
-  abstract adapt(source: S): CmsStructureModel;
-}
