+++ projects/core/src/cms/services/index.ts
@@ -1,6 +1,4 @@
 export * from './component-mapper.service';
-export * from './cms-page.loader';
-export * from './cms-page.adapter';
 export * from './cms-structure-config.service';
 export * from './component-mapper.service';
 export * from './dynamic-attribute.service';
